+++ tests/DocumentValidationTest/Validator/MyOwnEqualsValidator.php
@@ -2,9 +2,11 @@
 
 namespace Sokil\Mongo\DocumentValidationTest\Validator;
 
+use Sokil\Mongo\Structure;
+
 class MyOwnEqualsValidator extends \Sokil\Mongo\Validator
 {
-    public function validateField(\Sokil\Mongo\Document $document, $fieldName, array $params)
+    public function validateField(Structure $document, $fieldName, array $params)
     {
         if (!$document->get($fieldName)) {
             return;